+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 # Release 4.1.0 (?)
+## Bug fixes
+- Fix broken order by (#2638)
 
 ## Bug fixes
 - Report the Inmanta OSS product version correctly (#2622)