+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# Release 4.0.1 (?)
+# Release 4.1.0 (?)
 
 ## Bug fixes
 - Fix broken order by (#2638)